move print queue into formatting module
